db/pg: separate file 'models.go'
