+++ internal/db/pg/day.go
@@ -8,7 +8,45 @@ import (
 	"github.com/pkg/errors"
 
 	db_common "github.com/ShoshinNikita/budget-manager/internal/db"
+	"github.com/ShoshinNikita/budget-manager/internal/pkg/money"
 )
+
+// Day represents day entity in PostgreSQL db
+type Day struct {
+	tableName struct{} `pg:"days"` // nolint:structcheck,unused
+
+	// MonthID is a foreign key to 'months' table
+	MonthID uint `pg:"month_id"`
+
+	ID uint `pg:"id,pk"`
+
+	Day int `pg:"day"`
+	// Saldo is a DailyBudget - Cost of all Spends multiplied by 100 (can be negative)
+	Saldo  money.Money `pg:"saldo,use_zero"`
+	Spends []*Spend    `pg:"fk:day_id"`
+}
+
+// ToCommon converts Day to common Day structure from
+// "github.com/ShoshinNikita/budget-manager/internal/db" package
+func (d *Day) ToCommon(year int, month time.Month) *db_common.Day {
+	if d == nil {
+		return nil
+	}
+	return &db_common.Day{
+		ID:    d.ID,
+		Year:  year,
+		Month: month,
+		Day:   d.Day,
+		Saldo: d.Saldo,
+		Spends: func() []*db_common.Spend {
+			spends := make([]*db_common.Spend, 0, len(d.Spends))
+			for i := range d.Spends {
+				spends = append(spends, d.Spends[i].ToCommon(year, month, d.Day))
+			}
+			return spends
+		}(),
+	}
+}
 
 func (db DB) GetDay(_ context.Context, id uint) (*db_common.Day, error) {
 	var (
